psf-scale-factor: considering a background value in the scaling of the PSF
Until now, the PSF flux scaling factor was obtained just by considering the
ratio between the PSF and the star. In this situation, this ratio may be
overestimated because multiple reasons. In particular, one reason is that
the star could be on top of other (large and diffuse) source (e.g., galaxy
halo, stellar stream, etc.). As a consequence, the flux that does not
correspond to the star but to the contaminant source, was also taken into
account in the flux scaling factor. This led to an over estimation of the
flux and thus, a final over-subtraction of the stars. There is one solution
that is described in Infante-Sainz et al. 2020 (Section 2.4 PSF junction
(central-intermediate-outer parts), and Equation 5). The method there is
used for joining two PSFs at a given radius, but the method is the same for
modeling a star by the PSF.

With this commit, I have included some work to add the proposed solution as
part of this script. Now, the script will compute automatically the two
parameters (flux factor and background value) in order to better estimate
the flux calibration of the star. Still, more work is necessary but the
basis of this method is now available.