psf-scale-factor: improving the efficiency of the background estimation
Until now, the work done for estimating the background component was done
by using two nested loops and reading the radial profiles each time the
different values were needed. This was very unefficient because it implied
the reading of the datasets many different times. It is more efficient to
read the tables in advance and then access the data by indexes withing the
loops.

With this commit, the estimation of the background component has been
improved by including the solution explained above. Now the two profiles
are kept in shell arrays and thus the process is much more efficient than
before. In addition to this, other parts of the script have been modified
in order to include the necessary parameters and information.