Book: adding new information to the PSF subtraction tutorial
Until now, there was some missing information in the PSF subtraction
tutorial about the new features added to the PSF scripts.

With this commit, the tutorial describing how to model and subtract the PSF
has been updated. Now it includes information on how to use the new
features. Also, a bug in 'psf-scale-factor' that I found when writing the
tutorial has been corrected. It was about the injection of the fake
astrometry into the PSF image.