Book: adding the new PSF script features into the options information
Until now, I updated the PSF scripts and the tutorial in order to describe
the new features. However, some important information was still missing in
the actual information of the scripts.

With this commit, that information has been added to the part of the Book
in which the scripts are described.