radial-profile: --azimuth can be called multiple times
Until now, it was only possible to ask for a single azimuthal range when
generating the radial profile.

With this commit, it is now possible to call the '--azimuth' option
multiple times to cover azimuthally different/discrete regions of the
profile.

This commit is the result of rebasing a series of commits by Ignacio Ruiz
Cejudo who first implemented this feature. They were followed by edits by
Raul Infante-Sainz and a final review/edit by Mohammad Akhlaghi.