developer-build: fixed silent exits due to typos
Until now, due to a typo (forgetting the '$' sign behind variable names),
the script ended silently in case of a lack of parameters given to
'--top-build-dir', '--jobs' and '--compression'.

With this commit, the typos have been corrected and the user is notified
with an error message if they are given an empty value.

This fixes bug #65672