Arithmetic: no extra FITS file created when output is not FITS
Until now, when the output of Arithmetic was a plain-text file or the
standard output, Arithmetic would also build an empty (only 0-th HDU with
metadata) file! This happened because the step to write the metadata in the
0th HDU was always executed, even when the output was not a FITS file.

With this commit, a check has been added to only write the metadata HDU
when there actually is an output FITS file.

This bug was reported by Sepideh Eskandarlou.

This fixes bug #65673.