Library: OpenCL Integrated with Gnuastro
Untill now, Gnuastro had no support for GPU acceleration
and relied on pthreads for multithreaded operations.

With this commit, Gnuastro now has
cl-utils: A new module for interacting with OpenCL
through Gnuastro.

astconvolve: A new implementation of Convolution on the
GPU using OpenCL that is close to the previous pthread
implementation(which is now removed).