psf-scale-factor: minor stylistic polishing
Until now, there were some TABs within parts of this script, and some of
the pipes were not broken into a second line. Also within the
documentation, the two '@itemx's should be '@item'.

With this commit, to unify the format of the script code with the rest of
the scripts, they have been corrected; and the '@item's have also been
fixed.