Library (pdf.h): using fork+execl to return control back to Gnuastro
Until now, we were calling Ghostscript through the 'execl' function (which
was itself a recent replace from 'system()' in Commit 61fbd358). But
through 'execl' Gnuastro gives its full control to Ghostscript; not
allowing Gnuastro to clean up afterwards (in particular, the EPS file!).

With this commit, 'execl' is now only called on a child process of
Gnuastro. Therefore it only closes the child, not the parent and once it is
finished, Gnuastro is able to complete its cleaning up.

This issue (that the EPS file was not getting deleted was reported by
Sepideh Eskandarlou.