psf-stamp: corrected recently warping in label images for masking
Until now, we warped the image, clumps and objects map. But, when we wrap
it the clumps and objects labeled images (integers), they are changed to
floating point. Due to this the main object is masked in final output
(which is not what we expect).

With this commit, first we create a mask map of whole the images for the
desired clump+object. The output is therefore binary. Then we warp the
binary image and mask all non zero objects. This solves the problem of
directly warping integer labels, allowing proper masking of extra
clumps+objects while keeping the main object+clump.