Library (speclines.h): correct rounding of lines in thier name
Until now, the names of spectral lines that included their wavelength in
them were rounded to the lowest near integer (for example 'O-III-5006',
while the wavelength is 5006.84 Angstroms). The correct way to round is to
round by the nearest integer, not the lowest near integer!

With this commit, all the problematic macros and string that had such
wrongly rounded names have been corrected. Also, all wavelength values are
written with a fixed precision of three decimal digits.