astscript-radial-profile: removed bash-specific redirection
Until now, the radial profile script used the '&>' construct to re-direct
standard and error strings to '/dev/null'. However, this way of redirecting
is specific to Bash and may not work on other shells.

With this commit, the problem is fixed by replacing it with this command:
'> /dev/null 2>&1'.

This problem was reported by Phil Wyett, from the output of Debian's
Lintian.