Book: Re-write of the Brightness flux and magnitude section
Until now, this important section hadn't defined the important concept of
spectral flux density and incorrectly used "Brightness" (erg/s) in its
definition of Magnitudes. This has no effects on any of the programs, just
in terms of proper definitions. Furthermore, the derivation of the
conversion between Janskys and wavelength flux density was only in the
comments of the code (and thus not visible to a regular user).

With this commit, the section was fully reviewed and re-written to first
describe the physical concepts (brightness, flux, luminosity and spectral
flux density), then go into the astronomy-specific terminology of
magnitudes, zero point and surface brightness. Also, the derivation of the
conversion between Janskys and wavelength flux density was added to the
operator and the inverse operator was also defined.