Book: further edits on brightness, flux and magnitude section
Until now, the various important physical concepts (and their units) were
blended into the paragraph texts, making them hard to notice for a fast
access to one definition. However, later for the optical astronomy
terminology, we had divided each in a separate "table".

With this commit, the physical concepts are also discussed in a similar
format to the optical astronomy terms. The text was also edited to fit into
this structure.