MakeProfiles: first pixel of profile not checked for truncation
Until now, all the pixels of a profile where checked for being within the
truncation radius. As a result, when the given profile was extremely small
(such that its floating-point center would be more distant from the pixel's
center than the truncation radius) MakeProfiles would not produce any
pixels at all, returning NaN for all the profiles pixels!

With this commit, the distance of the pixel center from the profile center
(to check for trunction) is only applied after the first pixel. This solves
the problem described above.

In the debugging process, I noticed that one of the parts of the same
function (comment and description of the call to 'print_tossll') is no
longer applicable/useful, so I removed it.

This bug was reported by Alejandro Camazón Pinilla.

This fixes bug #66216.