MakeProfiles: Num. used threads printed if fewer profs than threads
Until now, MakeProfiles would always print the maximum number of available
threads in the system, even if it didn't use them. This could cause
confusion for advanced users who check their performance when the number of
input profiles was less than the total number of threads. Because
MakeProfiles builds one profile at each moment with each thread, in such
cases, it wouldn't use the full number of threads that were printed.

With this commit, the problem was solved by printing some extra
information. Also, when the total number of threads is more than the number
of profiles, it reminds the user that one thread is used for each profile.

This task was suggested by Carlos Marrero de La Rosa.