Convolve: check images written to directory of output, not input
Until now, when a check image was requested of Convolve, it was created in
the running directory, and based on the name of the input, not in the
output directory and based on the name of the output (as is expected in all
Gnuastro programs).

With this commit, this un-expected behavior is fixed. Also, in the
description of Convolve's '--makekernel' option, we now directly point to
the tutorial on building the extended PSF and the old list of how to use a
bright star to build the PSF is removed.

This fixes bug #66280.